Parse numbers, change string API, add more docs, and more
Move core data structures into Parser.Internal, and move low-level
state info into Parser.LowLevel.

Introduce the `int` and `float` parsers. They use backing code in
Parser.Internal.

Introduce `end` which lets you know when you are at the end of the
source string.

Replace `string` with `symbol` and `keyword`. That way (1) there is no
ambiguity about whether it is a parser for double-quoted strings and
(2) the error messages for symbol and keyword are more specific.