Optimize number parser
Only pass over base 2, 8, 10, and 16 numbers once. Do not allocate a
substring and then allocate each char to get the final number.

Instead, each of these computes the final number as it goes. So it will
now produce a tuple. The downside is that if you try to parse an
integer when there is not one around, you get the tuple allocation. The
old version would have allocated a Char here, so maybe it is not a bad
trade.