Update examples/elm.json to be ready for Elm 0.19.1
Before it was setup to be ready for Elm 0.19.0 which means
you see an error when trying to run the exmaple in Elm 0.19.1.

Now that Elm 0.19.1 is released it seems, to me at least, to be
a good idea for the examples to run by default on that version,
and for the user to have to modify it to run a different version.